# docs/openapi.yaml
openapi: 3.0.2

info:
  version: "1.0.0"
  title: WeatherMe
  description: Front-end website that show's temperature and humidity levels in a specific room to the user client.
  
servers:

  - description: Localhost
    url: http://127.0.0.1:3000/api/v1
    
paths: 
  /temperature:
      get:
        summary: get temperature
        description: fetch temperature data from mongodb
        operationId: GetTemperature
        security:
          - basicAuth: []
        responses:
          '200':
            description: temperature sensor valid response
            content:
              application/json:
                schema:
                  type: object
                  $ref: '#/components/schemas/temperature'
          '401':
            $ref: '#/components/responses/UnauthorizedError'
                 
      post:
        summary: post temperature
        description: post temperature data from sensors
        operationId: PostTemperature
        security:
          - basicAuth: []
        requestBody:
          description: temperature object
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                  location:
                    type: string
                  date:
                    type: integer
                  temperature:
                    type: integer
              example:
                  location: '1231'
                  date: 124124
                  temperature: 12312
        responses:
          '201':
            description: post temperature sucess response
            content:
              application/json:
                schema:
                  type: object
                  $ref: '#/components/schemas/temperature'
                example:
                  location: casa
                  date: 123456
                  temperature: 1234
          '401':
            $ref: '#/components/responses/UnauthorizedError'   
  /temperature/{id}:
      patch:
        summary: Patch temperature
        description: Update data
        operationId: updateTemperature
        security:
          - basicAuth: []
        parameters:
          - in: path 
            name: id
            schema:
              type: string
            required: true
        requestBody:
          description: temperature object
          required: true
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: integer
                  location:
                    type: string
                  date:                    
                    type: integer
                  temperature:
                    type: integer
              example:
                  location: '123'
                  date: 124124
                  temperature: 12312
        responses:
          '200':
            description: Temperature updated.
            content:
              application/json:
                schema:
                  type: object
                  $ref: '#/components/schemas/temperature'
                example:
                  location: campoUpdate
                  date: 989796
                  temperature: 55


components:
  securitySchemes:
    basicAuth:
      type: http  
      description: basic authentication   
      scheme: basic
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
  schemas:
    temperature:
      type: object
      required:
        - id
        - location
        - date
        - temperature
      properties:
        id:
          type: integer
        location:
          type: string
        date:
          type: integer
        temperature:
          type: integer
      example:
          - id: 123
      description: Type of TemperatureSensor object

security:
  - securityRequirement: []